requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==4.1
djangorestframework==3.14.0
dotenv==0.9.9
gunicorn==21.2.0
idna==3.11
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
xmltodict==0.13.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cron==0.6.0
